warnings except 'unused-function' and 'type-limits' clear
unused-function: Not cleared because it complains about
'gen_std_allocator()' like functions

type-limits: This is only in case of gcc, not clang
Case 1: Not cleared because in cabse of usage of 'for_range', most
of the times it is used to iterate from 0 upto some value, and the
data type used is unsigned type. Since to determine the direction of
update needed, we compare a < b for [a, b), and it will always be true
when a = 0.
Case 2: In case of slice_inx() macro, we also check if index is >=
0. This was done for ensuring that macro works for other signed types
also, but for unsigned types, just as above it gives out warnings.