Split slice definition macro for a reason
Since originally the DEF_SLICE() macro defined both the structure and
the function and the functions used sizeof() of the base type, you
couldnt include a slice of a type inside the type
itself,
which is a common usecase in creating n-ary trees.

So in this commit, Ive split the DEF_SLICE() into two macros
DEF_SLICE_STR() that defines the struct for the slice and
DEF_SLICE_FXN() that defines the functions to use with the slice

So whenever this special use case is needed, then the parts of the
DEF_SLICE() can be used. For other cases the whole of the definition
can be used directly.